+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 * ProfileAssociator Component lists are empty (UIDATIMP-399)
 * Fix UI tests (UIDATIMP-399)
 * Fix action profiles' sequence in job profiles (UIDATIMP-412)
+* Wipe out linked profiles when duplicate a profile (UIDATIMP-410)
 
 ## [1.7.3](https://github.com/folio-org/ui-data-import/tree/v1.7.3) (2019-12-04)
 * Update sorting query for jobs (UIDATIMP-346)